update tests that rely on model running
