+++ src/app/src/test/kotlin/org/imperial/mrc/hint/integration/DownloadTests.kt
@@ -8,25 +8,14 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.EnumSource
 import org.springframework.boot.test.web.client.getForEntity
 import org.springframework.boot.test.web.client.postForEntity
-import org.springframework.http.HttpEntity
-import org.springframework.http.HttpHeaders
-import org.springframework.http.MediaType
 import org.springframework.http.ResponseEntity
 
 class DownloadTests : SecureIntegrationTests() {
 
-    private fun getModelRunJsonEntity(): HttpEntity<String> {
-        val options = mapOf("sleep" to 1)
-        val headers = HttpHeaders()
-        headers.contentType = MediaType.APPLICATION_JSON
-        val jsonString = ObjectMapper().writeValueAsString(options)
-        return HttpEntity(jsonString, headers)
-    }
-
     private fun waitForModelRunResult(isAuthorized: IsAuthorized) : String {
         val id = when (isAuthorized) {
             IsAuthorized.TRUE -> {
-                val entity = getModelRunJsonEntity()
+                val entity = getModelRunEntity(isAuthorized)
                 val runResult = testRestTemplate.postForEntity<String>("/model/run/", entity)
                 ObjectMapper().readValue<JsonNode>(runResult.body!!)["data"]["id"].textValue()
             }
